LanguageMirror/LanguageMirror/2025-09-13/scripts/requirements.txt
openai-whisper
openai
python-dotenv
numpy
scipy
soundfile

